requirements.txt
requests
beautifulsoup4
google-generativeai
python-dotenv
tqdm
